+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Change Log
 All notable changes to this project will be documented in this file. This change log follows the conventions of [keepachangelog.com](http://keepachangelog.com/).
 
+# Unreleased
+## Added
+- navigation: `prepostwalk` & `prepostwalk-demo`
+
 # [0.13.2] - 2017-10-29
 ## Added
 - `shuriken.dev` namespace with `shuriken.monkey-patches.syntax-quote`.